Move enrollment locks into enrollment library
