+++ apps/prairielearn/src/lib/enrollment/lock.sql
