+++ apps/prairielearn/src/lib/enrollment/lock.ts
